Verificaciones por Admin y cliente
Las verificaciones dependiendo del rol del usuario

+++ src/controllers/categorias.controller.js
@@ -3,10 +3,16 @@ const Productos = require('../models/productos.model.js');
 
 //  OBTENER TODAS LAS CATEGORIAS
 function ObtenerCategorias(req, res) {
-    Categorias.find({}, (err, todasLasCategorias) => {
 
-        return res.send({ categorias: todasLasCategorias })
-    })
+    if (req.user.rol == "ADMIN") {
+        Categorias.find({}, (err, todasLasCategorias) => {
+
+            return res.send({ categorias: todasLasCategorias })
+        })
+    } else {
+        return res.send({ mensaje: "No está autorizado para realizar está accion" })
+    }
+    
 }
 
 // AGREGAR UNA CATEGORIA
@@ -14,16 +20,21 @@ function AgregarCategoria(req, res) {
     var parametros = req.body;
     var modCategoria = new Categorias();
 
-    if (parametros.categoria) {
-        modCategoria.categoria = parametros.categoria;
-
-        modCategoria.save((err, categoriaGuardada) => {
-
-            return res.send({ categoria: categoriaGuardada });
-        });
-    } else {
-        return res.send({ mensaje: "Debe de que enviar un nombre de Categoria" })
+     if(req.user.rol == "ADMIN"){
+        if (parametros.categoria) {
+            modCategoria.categoria = parametros.categoria;
+    
+            modCategoria.save((err, categoriaGuardada) => {
+    
+                return res.send({ categoria: categoriaGuardada });
+            });
+        } else {
+            return res.send({ mensaje: "Debe de que enviar un nombre de Categoria" })
+        }
+    }else{
+        return res.send({ mensaje: "No está autorizado para realizar está accion" })
     }
+
 }
 
 // EDITAR UNA CATEGORIA
@@ -31,26 +42,37 @@ function EditarCategoria(req, res) {
     var idCat = req.params.idCategoria;
     var parametros = req.body;
 
-    Categorias.findByIdAndUpdate(idCat, parametros, { new: true }, (err, categoriaEditada) => {
-        if (err) return res.status(500).send({ mensaje: 'Error en la peticion' });
-        if (!categoriaEditada) return res.status(404)
-            .send({ mensaje: 'Error al Editar la Categoria' });
-
-        return res.status(200).send({ categoria: categoriaEditada });
-    })
+    if (req.user.rol == "ADMIN") {
+        Categorias.findByIdAndUpdate(idCat, parametros, { new: true }, (err, categoriaEditada) => {
+            if (err) return res.status(500).send({ mensaje: 'Error en la peticion' });
+            if (!categoriaEditada) return res.status(404)
+                .send({ mensaje: 'Error al Editar la Categoria' });
+    
+            return res.status(200).send({ categoria: categoriaEditada });
+        })
+    }else{
+        return res.send({ mensaje: "No está autorizado para realizar está accion" })
+    }
+    
 }
 
 // ELIMINAR UNA CATEGORIA
 function EliminarCategoria(req, res) {
     var idCat = req.params.idCategoria;
 
-    Categorias.findByIdAndDelete(idCat, (err, categoriaEliminada) => {
-        if (err) return res.status(500).send({ mensaje: 'Error en la peticion' });
-        if (!categoriaEliminada) return res.status(500)
-            .send({ mensaje: 'Error al eliminar la Categoria' })
 
-        return res.status(200).send({ categoria: categoriaEliminada });
-    })
+    if (req.user.rol == "ADMIN") {
+        Categorias.findByIdAndDelete(idCat, (err, categoriaEliminada) => {
+            if (err) return res.status(500).send({ mensaje: 'Error en la peticion' });
+            if (!categoriaEliminada) return res.status(500)
+                .send({ mensaje: 'Error al eliminar la Categoria' })
+    
+            return res.status(200).send({ categoria: categoriaEliminada });
+        })
+    } else {
+        return res.send({ mensaje: "No está autorizado para realizar está accion" })
+    }
+    
 }
 
 function crearCategoriaAlIniciar(req, res) {
